requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
